Split Room class in a file
