+++ server.py
@@ -1,6 +1,8 @@
 from flask import Flask
 from flask_socketio import SocketIO, emit
 
+from room import Room
+
 app = Flask(__name__)
 socketio = SocketIO(app)
 
@@ -9,51 +11,29 @@ app.debug = True
 participants = []
 
 
-class Room:
-    def __init__(self):
-        self.players = []
-
-    def join(self, name):
-        player = {'name': name, 'choice': None}
-        self.players.append(player)
-        self.broadcast_players()
-
-    def leave(self, name):
-        for player in self.players:
-            if player['name'] == name:
-                self.players.remove(player)
-                self.broadcast_players()
-                break
-
-    def select(self, name, choice):
-        for player in self.players:
-            if player['name'] == name:
-                player['choice'] = choice
-                self.broadcast_players()
-                break
-
-    def broadcast_players(self):
-        emit('players_update', self.players, broadcast=True)
-
-
 room = Room()
+
+
+def broadcast_players(room):
+    emit('players_update', room.players, broadcast=True)
 
 
 @socketio.on('join')
 def handle_join(name):
     room.join(name)
+    broadcast_players(room)
 
 
 @socketio.on('left')
 def handle_left(name):
     room.leave(name)
+    broadcast_players(room)
 
 
 @socketio.on('select')
 def handle_select(player):
     room.select(player['name'], player['choice'])
-
-
+    broadcast_players(room)
 
 
 @app.route('/')
